ISPN-5301 Make CompositeNotifyingFuture more generic
